+++ apps/dine-pleiepenger/src/pages/api/mellomlagring.api.ts
@@ -1,11 +1,10 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { createChildLogger } from '@navikt/next-logger';
 import { storageParser } from '@navikt/sif-common-core-ds/src/utils/persistence/storageParser';
 import axios from 'axios';
 import { withAuthenticatedApi } from '../../auth/withAuthentication';
 import { fetchMellomlagringer } from '../../server/apiService';
 import { Mellomlagringer } from '../../types/Mellomlagring';
-import { getXRequestId } from '../../utils/apiUtils';
+import { getLogger } from '../../utils/getLogCorrelationID';
 
 export const mellomlagringFetcher = async (url: string): Promise<Mellomlagringer> =>
     axios.get(url, { transformResponse: storageParser }).then((res) => res.data);
@@ -14,8 +13,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
     try {
         res.send(await fetchMellomlagringer(req));
     } catch (err) {
-        const childLogger = createChildLogger(getXRequestId(req));
-        childLogger.error(`Hent mellomlagring feilet: ${err}`);
+        getLogger(req).error(`Hent mellomlagring feilet: ${err}`);
         res.status(500).json({ error: 'Kunne ikke hente mellomlagring' });
     }
 }