+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/types/SøknadRoutes.ts
@@ -2,7 +2,8 @@ import { StepId } from './StepId';
 
 export const SøknadStepRoutePath = {
     [StepId.VELKOMMEN]: 'velkommen',
-    [StepId.FOSTERBARN]: 'fosterbarn',
+    [StepId.DINE_BARN]: 'dine-barn',
+    [StepId.DELT_BOSTED]: 'delt-bosted',
     [StepId.SITUASJON]: 'situasjon',
     [StepId.FRAVÆR]: 'fravær',
     [StepId.LEGEERKLÆRING]: 'legeerklaring',
@@ -14,7 +15,8 @@ export const SøknadStepRoutePath = {
 export enum SøknadRoutes {
     INNLOGGET_ROOT = '/soknad/*',
     VELKOMMEN = '/soknad/velkommen',
-    FOSTERBARN = '/soknad/fosterbarn',
+    DINE_BARN = '/soknad/dine-barn',
+    DELT_BOSTED = '/soknad/delt-bosted',
     SITUASJON = '/soknad/situasjon',
     FRAVÆR = '/soknad/fravær',
     LEGEERKLÆRING = '/soknad/legeerklaring',