+++ packages/sif-common-formik-ds/src/utils/eøsCountries.ts
@@ -1,0 +1,161 @@
+/** Kilde:
+ * https://www.nav.no/no/person/flere-tema/arbeid-og-opphold-i-utlandet/relatert-informasjon/eos-landene
+ */
+
+export const eøsCountries = [
+    {
+        alpha2: 'BE',
+        alpha3: 'BEL',
+        name: 'Belgia',
+    },
+    {
+        alpha2: 'BG',
+        alpha3: 'BGR',
+        name: 'Bulgaria',
+    },
+    {
+        alpha2: 'DK',
+        alpha3: 'DNK',
+        name: 'Danmark',
+    },
+    {
+        alpha2: 'EE',
+        alpha3: 'EST',
+        name: 'Estland',
+    },
+    {
+        alpha2: 'FI',
+        alpha3: 'FIN',
+        name: 'Finland',
+    },
+    {
+        alpha2: 'FR',
+        alpha3: 'FRA',
+        name: 'Frankrike',
+    },
+    {
+        alpha2: 'GR',
+        alpha3: 'GRC',
+        name: 'Hellas',
+    },
+    {
+        alpha2: 'IE',
+        alpha3: 'IRL',
+        name: 'Irland',
+    },
+    {
+        alpha2: 'IS',
+        alpha3: 'ISL',
+        name: 'Island',
+    },
+    {
+        alpha2: 'IT',
+        alpha3: 'ITA',
+        name: 'Italia',
+    },
+    {
+        alpha2: 'HR',
+        alpha3: 'HRV',
+        name: 'Kroatia',
+    },
+    {
+        alpha2: 'CY',
+        alpha3: 'CYP',
+        name: 'Kypros',
+    },
+    {
+        alpha2: 'LV',
+        alpha3: 'LVA',
+        name: 'Latvia',
+    },
+    {
+        alpha2: 'LI',
+        alpha3: 'LIE',
+        name: 'Liechtenstein',
+    },
+    {
+        alpha2: 'LT',
+        alpha3: 'LTU',
+        name: 'Litauen',
+    },
+    {
+        alpha2: 'LU',
+        alpha3: 'LUX',
+        name: 'Luxembourg',
+    },
+    {
+        alpha2: 'MT',
+        alpha3: 'MLT',
+        name: 'Malta',
+    },
+    {
+        alpha2: 'NL',
+        alpha3: 'NLD',
+        name: 'Nederland',
+    },
+    {
+        alpha2: 'NO',
+        alpha3: 'NOR',
+        name: 'Norge',
+    },
+    {
+        alpha2: 'PL',
+        alpha3: 'POL',
+        name: 'Polen',
+    },
+    {
+        alpha2: 'PT',
+        alpha3: 'PRT',
+        name: 'Portugal',
+    },
+    {
+        alpha2: 'RO',
+        alpha3: 'ROU',
+        name: 'Romania',
+    },
+    {
+        alpha2: 'SK',
+        alpha3: 'SVK',
+        name: 'Slovakia',
+    },
+    {
+        alpha2: 'SI',
+        alpha3: 'SVN',
+        name: 'Slovenia',
+    },
+    {
+        alpha2: 'ES',
+        alpha3: 'ESP',
+        name: 'Spania',
+    },
+    {
+        alpha2: 'CH',
+        alpha3: 'CHE',
+        name: 'Sveits',
+    },
+    {
+        alpha2: 'SE',
+        alpha3: 'SWE',
+        name: 'Sverige',
+    },
+    {
+        alpha2: 'CZ',
+        alpha3: 'CZE',
+        name: 'Tsjekkia',
+    },
+    {
+        alpha2: 'DE',
+        alpha3: 'DEU',
+        name: 'Tyskland',
+    },
+    {
+        alpha2: 'HU',
+        alpha3: 'HUN',
+        name: 'Ungarn',
+    },
+    {
+        alpha2: 'AT',
+        alpha3: 'AUT',
+        name: 'Østerrike',
+    },
+];