+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @navikt/omsorgspengerutbetaling-arbeidstaker-soknad
 
+## 1.3.5
+
+### Patch Changes
+
+-   99e1bd2: Pakkeoppdateringer - minor/patch
+-   Updated dependencies [99e1bd2]
+    -   @navikt/sif-common-formik-ds@1.9.1
+    -   @navikt/sif-common-soknad-ds@10.15.9
+    -   @navikt/sif-common-forms-ds@8.23.1
+    -   @navikt/sif-common-core-ds@9.17.10
+    -   @navikt/sif-common-sentry@0.26.7
+    -   @navikt/sif-common-amplitude@2.25.5
+    -   @navikt/sif-common-utils@3.46.3
+
+## 1.3.4
+
+### Patch Changes
+
+-   Updated dependencies [a739a56]
+    -   @navikt/sif-common-formik-ds@1.9.0
+    -   @navikt/sif-common-forms-ds@8.23.0
+    -   @navikt/sif-common-core-ds@9.17.9
+    -   @navikt/sif-common-soknad-ds@10.15.8
+
 ## 1.3.3
 
 ### Patch Changes